refactor: Consolidated s3.ts files
