+++ src/data/s3.ts
@@ -1,6 +1,11 @@
 import { S3Event } from 'aws-lambda';
 import AWS from 'aws-sdk';
 
+export interface S3ObjectParameters {
+    Bucket: string;
+    Key: string;
+}
+
 const getS3Client = (): AWS.S3 => {
     let options = {};
 
@@ -17,22 +22,30 @@ const getS3Client = (): AWS.S3 => {
 };
 
 const s3 = getS3Client();
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const fetchDataFromS3 = async (event: S3Event): Promise<any> => {
+
+export const getFileFromS3 = async (params: S3ObjectParameters): Promise<string> => {
+    const data = await s3.getObject(params).promise();
+    return data.Body?.toString('utf-8') ?? '';
+};
+
+export const fetchDataFromS3 = async <T>(event: S3Event, isEmailer = false): Promise<T> => {
     try {
-        const s3BucketName: string = event.Records[0].s3.bucket.name;
-        const s3FileName: string = decodeURIComponent(event.Records[0].s3.object.key.replace(/\+/g, ' '));
-        const dataAsString: string =
-            (
-                await s3
-                    .getObject({
-                        Bucket: s3BucketName,
-                        Key: s3FileName,
-                    })
-                    .promise()
-            ).Body?.toString('utf-8') ?? '';
-        const dataAsJson: JSON = JSON.parse(dataAsString);
-        return dataAsJson;
+        const s3BucketName: string = !isEmailer
+            ? event.Records[0].s3.bucket.name
+            : process.env.MATCHING_DATA_BUCKET || '';
+
+        const s3FileName: string = !isEmailer
+            ? decodeURIComponent(event.Records[0].s3.object.key.replace(/\+/g, ' '))
+            : event.Records[0].s3.object.key.replace('.xml', '.json');
+
+        const params: S3ObjectParameters = {
+            Bucket: s3BucketName,
+            Key: s3FileName,
+        };
+
+        const dataAsString: string = await getFileFromS3(params);
+
+        return JSON.parse(dataAsString);
     } catch (err) {
         throw new Error(`Error in retrieving data. ${err.stack}`);
     }
